Ignore vscode history directory when running tests
The local history plugin in vscode creates a .history directory in the
project workspace (by default, though it can be configured). Jest needs
to be told to ignore this directory otherwise it will try to run any
tests it find in there (which it will if you are editing tests).